Rename psi:pack into psi:iconpack
