+++ components/psi/iconpack/src/main/kotlin/io/github/composegears/valkyrie/psi/iconpack/IconPackPsiParser.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.psi.pack
+package io.github.composegears.valkyrie.psi.iconpack
 
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiManager
