+++ components/psi/iconpack/src/test/resources/DataObjectIconPack.kt
