+++ components/psi/iconpack/src/test/resources/NestedIconPack.kt
