+++ components/psi/iconpack/src/test/resources/SimpleIconPack.kt
